Move handleFileDownload back
